requirements.txt
# 核心依赖
streamlit>=1.28.0
pandas>=2.0.0

# 数据可视化
plotly>=5.18.0
matplotlib>=3.7.0
wordcloud>=1.9.0
pillow>=10.0.0

#文本分析
jieba>=0.42.1
snownlp>=0.12.3

# 数据处理
numpy>=1.24.0
openpyxl>=3.1.0

# HTTP请求
requests>=2.31.0

# 异步HTTP请求
aiohttp>=3.8.0

# OpenAI API
openai>=1.3.0
python-dotenv>=1.0.0
Flask==2.0.1
requests==2.26.0
numpy==1.21.2